refactor(leetcode): 优化路径总和计算代码
- 移除了不必要的类型转换，将 (long) targetSum 改为 targetSum
- 这个改动简化了代码，提高了可读性
- 经过测试，这个改动不会影响功能，但可能略微提高性能

+++ src/main/java/com/so/leetcode/Q437_PathSumIII.java
@@ -38,7 +38,7 @@ public class Q437_PathSumIII {
         }
 
         // 从根节点开始计算路径和
-        countPaths(root, (long) targetSum);
+        countPaths(root, targetSum);
 
         // 对左右子树递归计算路径和
         pathSum(root.left, targetSum);
